requirements.txt
tensorflow-cpu>=2.6.0
fastapi>=0.68.0
uvicorn>=0.15.0
pillow>=8.3.1
numpy>=1.21.0
python-multipart>=0.0.5